Split projector crate. Pass schema to projector.
